pyproject.toml
[tool.poetry]
name = "try-drf"
version = "0.1.0"
description = ""
authors = ["oleksandr.y <oleksandr.yurchenko@orderbird.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.5"
algoliasearch = "^3.0.0"
algoliasearch-django = "^3.0.0"
jinja2 = "^3.1.4"
pygments = "^2.18.0"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
